+++ Client/Frontend/Login Management/LoginListViewController.swift
@@ -159,7 +159,7 @@ class LoginListViewController: SensitiveViewController {
         applyTheme()
 
         KeyboardHelper.defaultHelper.addDelegate(self)
-        viewModel.dataSource.delegate = self
+        viewModel.delegate = self
     }
 
     override func viewWillAppear(_ animated: Bool) {
@@ -227,7 +227,7 @@ class LoginListViewController: SensitiveViewController {
     }
 
     fileprivate func toggleSelectionTitle() {
-        if loginSelectionController.selectedCount == viewModel.dataSource.count {
+        if loginSelectionController.selectedCount == viewModel.count {
             selectionButton.setTitle(deselectAllTitle, for: [])
         } else {
             selectionButton.setTitle(selectAllTitle, for: [])
@@ -297,7 +297,7 @@ private extension LoginListViewController {
             self.deleteAlert = UIAlertController.deleteLoginAlertWithDeleteCallback({ [unowned self] _ in
                 // Delete here
                 let guidsToDelete = self.loginSelectionController.selectedIndexPaths.compactMap { indexPath in
-                    self.viewModel.dataSource.loginAtIndexPath(indexPath)?.id
+                    self.viewModel.loginAtIndexPath(indexPath)?.id
                 }
 
                 self.viewModel.profile.logins.delete(ids: guidsToDelete).uponQueue(.main) { _ in
@@ -312,7 +312,7 @@ private extension LoginListViewController {
 
     @objc func tappedSelectionButton() {
         // If we haven't selected everything yet, select all
-        if loginSelectionController.selectedCount < viewModel.dataSource.count {
+        if loginSelectionController.selectedCount < viewModel.count {
             // Find all unselected indexPaths
             let unselectedPaths = tableView.allLoginIndexPaths.filter { indexPath in
                 return !loginSelectionController.indexPathIsSelected(indexPath)
@@ -375,7 +375,7 @@ extension LoginListViewController: UITableViewDelegate {
             loginSelectionController.selectIndexPath(indexPath)
             toggleSelectionTitle()
             toggleDeleteBarButton()
-        } else if let login = viewModel.dataSource.loginAtIndexPath(indexPath) {
+        } else if let login = viewModel.loginAtIndexPath(indexPath) {
             tableView.deselectRow(at: indexPath, animated: true)
             let detailViewController = LoginDetailViewController(profile: viewModel.profile, login: login)
             detailViewController.settingsDelegate = settingsDelegate
@@ -431,12 +431,12 @@ extension LoginListViewController: SearchInputViewDelegate {
 }
 
 // MARK: - LoginViewModelDelegate
-extension LoginListViewController: LoginDataSourceViewModelDelegate {
+extension LoginListViewController: LoginViewModelDelegate {
 
     func loginSectionsDidUpdate() {
         loadingView.isHidden = true
         tableView.reloadData()
-        navigationItem.rightBarButtonItem?.isEnabled = viewModel.dataSource.count > 0
+        navigationItem.rightBarButtonItem?.isEnabled = viewModel.count > 0
         restoreSelectedRows()
     }
